requirements.txt
annotated-types==0.7.0
anyio==3.7.1
click==8.2.1
fastapi==0.104.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.24.0
